+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enh #88: Don't create cache directory on `FileCache` initialization (@vjik)
 - Bug #88: Set correct permissions for nested directories (@vjik)
 - Bug #85: Clear stat cache in `FileCache::set()` (@samdark)
+- Bug #86: Handle race condition when creating cache directory (@vjik)
 
 ## 3.1.0 October 09, 2023